refactor(discord): decompose recovery_engine.rs — completion_delivery + restore_inflight children (#3834)
Verbatim moves (python-verified against pre-move HEAD): recovery
completion/relay-notice cluster to completion_delivery.rs and the
restart-scan/session-retry cluster to restore_inflight.rs. Root drops
3571 -> 472 lines; restore_inflight.rs (2335 prod LoC) registered as a
tracked giant with baseline lowered 3039 -> 2335.

Relocated follow-up sites: #4111 whole-row save now at
restore_inflight.rs:1925-1928; #4117 session-died retry signal at
restore_inflight.rs:2260-2262.
